+++ packages/toolkit/config/webpack/devServer.js
@@ -1,12 +1,10 @@
-const path = require('path');
-
-module.exports = ({ projectConfig: { devServer, devServerPort } }) => {
-	if (!devServer) {
+module.exports = ({ isPackage, projectConfig: { devServer, devServerPort } }) => {
+	if (!devServer || !isPackage) {
 		return undefined;
 	}
 
 	return {
-		contentBase: path.join(__dirname, 'public'),
+		contentBase: 'public',
 		compress: true,
 		port: Number(devServerPort),
 	};